feat:统一将用户id的类型由uint64_t改为int64_t，同时修复联系人列表的bug，要区分不同分组的列表项，因为新的朋友分组和联系人分组中的列表项的用户id用户名完全一样，但是状态不同，所以要区分对待。

+++ app/appcore.h
@@ -40,7 +40,7 @@ public slots:
     void ReceiveSlotAddFriendReq(const QByteArray& data);
     void ReceiveSlotAddFriendAck(const QByteArray& data);
     void SendSlotAddFriendReqAck(const QByteArray& data, const User& user, FriendState type);
-    void SlotSendChatMsg(int id, const QString& content);
+    void SlotSendChatMsg(int64_t id, const QString& content);
     void SlotOneChat(const QByteArray& data);
 
 private slots:
